requirements-dev.txt
black==19.10b0
eth-brownie>=1.14.3,<2.0.0
python-dotenv==0.15.0